FEATBL-1437 feat: Make t_container_item.id 64 bit long
The upgrade works in two steps:
 - First, apply type migration on small tables, at normal upgrade time
 - Then, add a column "item_id64" on every "big" table, with a trigger
   updating item_id64 to match item_id if inserted or updated.
 - Finally, in a HotUpgrade, do the "long" stuff on big tables, trying to not block the database for long.

t_conversation is the only table which requires a somewhat "long" lock, while upgrading the foreign key to item_id.

+++ parent/calendar/net.bluemind.calendar.sqlschema/sql/calendarview-schema.1.0.0.sql
@@ -2,13 +2,12 @@ create type t_calendarview_type as enum
   ('DAY', 'WEEK', 'MONTH', 'LIST');
   
 create table t_calendarview (
-    item_id int4 references t_container_item(id) on delete cascade,
-
-    id  serial primary key,
-    label text,
-    type  t_calendarview_type,
-    calendars text[],
-    is_default boolean default false
+  item_id bigint references t_container_item(id) on delete cascade,
+  id  serial primary key,
+  label text,
+  type  t_calendarview_type,
+  calendars text[],
+  is_default boolean default false
 );
 
 create index idx_t_calendarview_item_id on t_calendarview (item_id);
