+++ parent/calendar/net.bluemind.calendar.sqlschema/sql/vevent-schema.1.0.0.sql
@@ -3,7 +3,6 @@ CREATE EXTENSION IF NOT EXISTS hstore WITH schema pg_catalog;
 create type t_calendar_transparency as enum ('Opaque', 'Transparent');
 
 create table t_calendar_vevent (
-
   dtstart_timestamp timestamp without time zone NOT NULL,
   dtstart_timezone timezone,
   dtstart_precision e_datetime_precision NOT NULL,
@@ -124,9 +123,9 @@ create table t_calendar_vevent (
    /*
     * 3.8.1.1.  Attachment
     */
-   attach_uri text[],
-   attach_name text[],
-   attach_cid text[],
+  attach_uri text[],
+  attach_name text[],
+  attach_cid text[],
 
   /*
    * 4.8.7.4.  Sequence Number
@@ -143,7 +142,7 @@ create table t_calendar_vevent (
   
   custom_properties hstore,
 
-  item_id int4 references t_container_item(id) on delete cascade
+  item_id bigint references t_container_item(id) on delete cascade
 );
 
 CREATE INDEX tcv_item_id_fkey ON t_calendar_vevent(item_id);
@@ -154,7 +153,7 @@ CREATE TABLE t_calendar_series (
   ics_uid text NOT NULL,
   properties hstore,
   accept_counters boolean,
-  item_id int4 REFERENCES t_container_item(id) ON UPDATE CASCADE ON DELETE CASCADE
+  item_id bigint REFERENCES t_container_item(id) ON UPDATE CASCADE ON DELETE CASCADE
 );
 CREATE INDEX t_calendar_series_item_id ON t_calendar_series(item_id);
 CREATE INDEX idx_calendar_series_lowercase_icsuid ON t_calendar_series (lower(ics_uid));
