+++ parent/core/net.bluemind.core.container.service.tests/sql/dummy-schema-1.0.0.sql
@@ -1,7 +1,7 @@
 create table t_dummy_value (
 	seen boolean default false,
 	deleted boolean default false,
-	item_id int4 references t_container_item(id) on delete cascade
+	item_id bigint references t_container_item(id) on delete cascade
 );
 
 create index dummyidx on t_dummy_value(item_id);
