+++ parent/core/net.bluemind.core.container.sqlschema/sql/containers-hierarchy-1.0.0-SNAPSHOT.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS  t_container_hierarchy (
 	name text not null,
 	container_type text not null,
 	container_uid text not null,
-	item_id integer not null references t_container_item(id) on delete cascade
+	item_id bigint not null references t_container_item(id) on delete cascade
 );
 CREATE INDEX IF NOT EXISTS idx_tch_item_id ON t_container_hierarchy(item_id);
 CREATE INDEX IF NOT EXISTS idx_tch_type  ON t_container_hierarchy USING gin(container_type);
