+++ parent/core/net.bluemind.core.container.sqlschema/sql/owner-subscriptions-1.0.0-SNAPSHOT.sql
@@ -5,6 +5,6 @@ CREATE TABLE IF NOT EXISTS t_owner_subscription (
 	owner text not null,
 	default_container boolean not null,
 	name text not null,
-	item_id integer not null references t_container_item(id) on delete cascade
+	item_id bigint not null references t_container_item(id) on delete cascade
 );
 CREATE INDEX IF NOT EXISTS idx_tos_item_id ON t_owner_subscription(item_id);
