+++ parent/user/net.bluemind.user.accounts.sqlschema/sql/user-accounts-schema.1.0.0.sql
@@ -1,9 +1,9 @@
 CREATE EXTENSION IF NOT EXISTS hstore WITH schema pg_catalog;
 
 create table t_user_account (
-	item_id 	  int4 references t_container_item(id) on delete cascade,
+	item_id 	  bigint references t_container_item(id) on delete cascade,
 	login 		  varchar(64) not null,
-	credentials 	  text,
+	credentials 	text,
 	system		  varchar(255),
 	properties hstore,
 	PRIMARY KEY (item_id, system)
